+++ ezfilter/src/main/java/cn/ezandroid/ezfilter/image/BitmapBuilder.java
@@ -1,0 +1,65 @@
+package cn.ezandroid.ezfilter.image;
+
+import android.graphics.Bitmap;
+
+import cn.ezandroid.ezfilter.EZFilter;
+import cn.ezandroid.ezfilter.core.FBORender;
+import cn.ezandroid.ezfilter.core.FilterRender;
+import cn.ezandroid.ezfilter.environment.IFitView;
+import cn.ezandroid.ezfilter.extra.IAdjustable;
+import cn.ezandroid.ezfilter.image.offscreen.OffscreenImage;
+
+/**
+ * 图片处理构造器
+ */
+public class BitmapBuilder extends EZFilter.Builder {
+
+    private Bitmap mBitmap;
+
+    public BitmapBuilder(Bitmap bitmap) {
+        mBitmap = bitmap;
+    }
+
+    public Bitmap output() {
+        // 离屏渲染
+        OffscreenImage offscreenImage = new OffscreenImage(mBitmap);
+        for (FilterRender filterRender : mFilterRenders) {
+            offscreenImage.addFilterRender(filterRender);
+        }
+        return offscreenImage.capture();
+    }
+
+    public Bitmap output(int width, int height) {
+        // 离屏渲染
+        OffscreenImage offscreenImage = new OffscreenImage(mBitmap);
+        for (FilterRender filterRender : mFilterRenders) {
+            offscreenImage.addFilterRender(filterRender);
+        }
+        return offscreenImage.capture(width, height);
+    }
+
+    @Override
+    protected FBORender getStartPointRender(IFitView view) {
+        return new BitmapInput(mBitmap);
+    }
+
+    @Override
+    protected float getAspectRatio(IFitView view) {
+        return mBitmap.getWidth() * 1.0f / mBitmap.getHeight();
+    }
+
+    @Override
+    public BitmapBuilder addFilter(FilterRender filterRender) {
+        return (BitmapBuilder) super.addFilter(filterRender);
+    }
+
+    @Override
+    public <T extends FilterRender & IAdjustable> BitmapBuilder addFilter(T filterRender, float progress) {
+        return (BitmapBuilder) super.addFilter(filterRender, progress);
+    }
+
+    @Override
+    public BitmapBuilder enableRecord(String outputPath, boolean recordVideo, boolean recordAudio) {
+        return (BitmapBuilder) super.enableRecord(outputPath, recordVideo, recordAudio);
+    }
+}
