+++ ezfilter/src/main/java/cn/ezandroid/ezfilter/view/ViewBuilder.java
@@ -1,0 +1,54 @@
+package cn.ezandroid.ezfilter.view;
+
+import cn.ezandroid.ezfilter.EZFilter;
+import cn.ezandroid.ezfilter.core.FBORender;
+import cn.ezandroid.ezfilter.core.FilterRender;
+import cn.ezandroid.ezfilter.core.RenderPipeline;
+import cn.ezandroid.ezfilter.environment.IFitView;
+import cn.ezandroid.ezfilter.extra.IAdjustable;
+import cn.ezandroid.ezfilter.view.glview.IGLView;
+
+/**
+ * 视图处理构造器
+ */
+public class ViewBuilder extends EZFilter.Builder {
+
+    private IGLView mGLView;
+
+    public ViewBuilder(IGLView view) {
+        mGLView = view;
+    }
+
+    @Override
+    protected FBORender getStartPointRender(IFitView view) {
+        return new ViewInput(mGLView);
+    }
+
+    @Override
+    protected float getAspectRatio(IFitView view) {
+        if (mGLView.getHeight() != 0) {
+            return mGLView.getWidth() * 1.0f / mGLView.getHeight();
+        }
+        return 1;
+    }
+
+    @Override
+    public ViewBuilder addFilter(FilterRender filterRender) {
+        return (ViewBuilder) super.addFilter(filterRender);
+    }
+
+    @Override
+    public <T extends FilterRender & IAdjustable> ViewBuilder addFilter(T filterRender, float progress) {
+        return (ViewBuilder) super.addFilter(filterRender, progress);
+    }
+
+    @Override
+    public ViewBuilder enableRecord(String outputPath, boolean recordVideo, boolean recordAudio) {
+        return (ViewBuilder) super.enableRecord(outputPath, recordVideo, recordAudio);
+    }
+
+    public RenderPipeline into(IFitView view) {
+        mGLView.setGLEnvironment(view);
+        return super.into(view);
+    }
+}
